superlingo_be/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
torch
TTS
numpy
soundfile